+++ src/main/main.ts
@@ -13,24 +13,16 @@ import 'core-js/stable';
 import 'regenerator-runtime/runtime';
 import path from 'path';
 import { app, BrowserWindow, shell, ipcMain, screen } from 'electron';
-import * as fs from 'fs';
+
 import log from 'electron-log';
 import MenuBuilder from './menu';
 import { resolveHtmlPath } from './util';
 import i18n from '../i18n/i18n';
-import {
-  dbPath,
-  dbUrl,
-  extraResourcesPath,
-  isDevelopment,
-  latestMigration,
-  Migration,
-} from './utils/defaultConstants';
+import { isDevelopment } from './utils/defaultConstants';
 import getAppDataFolder from './utils/fsUtils';
 import prismaClient from './database/prismaContext';
-import runPrismaCommand from './utils/runPrismaCommand';
-import seed from './database/seed';
 import logSystemConfigs from './utils/logSystemConfigs';
+import runDbMigrations from './database/migrationHandler';
 
 log.transports.file.resolvePath = () =>
   path.resolve(
@@ -95,62 +87,7 @@ const installExtensions = async () => {
 const createWindow = async () => {
   log.info('Creating a new window');
 
-  let needsMigration = false;
-  let mustSeed = false;
-  log.info(`Checking database at ${dbPath}`);
-  const dbExists = fs.existsSync(dbPath);
-
-  if (!dbExists) {
-    log.info('Database does not exists. Migration and seeding is needed.');
-    needsMigration = true;
-    mustSeed = true;
-    // since prisma has trouble if the database file does not exist, touches an empty file
-    log.info('Touching database file.');
-    fs.closeSync(fs.openSync(dbPath, 'w'));
-  } else {
-    log.info('Database exists. Verifying the latest migration');
-    try {
-      const latest: Migration[] =
-        await prismaClient.$queryRaw`select * from _prisma_migrations order by finished_at`;
-      log.info(
-        `Latest migration: ${latest[latest.length - 1]?.migration_name}`
-      );
-      needsMigration =
-        latest[latest.length - 1]?.migration_name !== latestMigration;
-    } catch (e) {
-      log.info('Latest migration could not be found, migration is needed');
-      log.error(e);
-      needsMigration = true;
-    }
-  }
-
-  if (needsMigration) {
-    try {
-      const schemaPath = isDevelopment
-        ? path.resolve(extraResourcesPath, 'prisma', 'schema.prisma')
-        : path.resolve(app.getAppPath(), '..', 'prisma', 'schema.prisma');
-      log.info(
-        `Database needs a migration. Running prisma migrate with schema path ${schemaPath}`
-      );
-
-      await runPrismaCommand({
-        command: ['migrate', 'deploy', '--schema', schemaPath],
-        dbUrl,
-      });
-
-      log.info('Migration done.');
-
-      if (mustSeed) {
-        await seed(prismaClient);
-      }
-    } catch (e) {
-      log.error('Migration executed with error.');
-      log.error(e);
-      process.exit(1);
-    }
-  } else {
-    log.info('Does not need migration');
-  }
+  await runDbMigrations();
 
   if (isDevelopment) {
     log.info('Installing additional dev extensions');
